refactor(db): modularize locality repository implementation
This commit refactors the `LocalityRepository` implementation by splitting its monolithic structure into smaller, more focused modules. The original `locality_repository_impl.rs` and `locality_repository_batch_impl.rs` files were becoming overly large and difficult to maintain. By breaking them down into individual files for each function (e.g., `find_by_id.rs`, `create_batch.rs`), the new structure improves code organization, readability, and maintainability.

This modular approach aligns with the existing pattern used for other repositories in the system, ensuring consistency across the codebase. The corresponding tests have been removed as they are now obsolete and will be replaced with new tests that reflect the updated structure.

Components affected:
- `banking-db-postgres`: Refactored `LocalityRepository` implementation
- `banking-db-postgres`: Removed obsolete test files for `LocalityRepository`
- `docs`: Updated progress tracking to reflect repository refactoring

Benefits:
- Improved Code Organization: Separates distinct functionalities into individual modules.
- Enhanced Readability: Smaller, focused files are easier to understand and navigate.
- Increased Maintainability: Simplifies future modifications and debugging.
- Consistent Architecture: Aligns the `LocalityRepository` with the established repository pattern.
