+++ banking-db-postgres/src/repository/person/locality_repository/repo_impl.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 use banking_api::BankingResult;
 use banking_db::models::person::{LocalityIdxModel, LocalityIdxModelCache, LocalityModel};
 use banking_db::repository::{
-    CountrySubdivisionRepository, LocalityRepository, LocalityRepositoryError, LocalityResult,
+    LocalityRepository, LocalityResult,
     TransactionAware,
 };
 use crate::repository::executor::Executor;
@@ -13,11 +13,9 @@ use once_cell::sync::OnceCell;
 use sqlx::{postgres::PgRow, Postgres, Row};
 use std::collections::{HashMap, HashSet};
 use std::error::Error;
-use std::hash::Hasher;
 use parking_lot::RwLock;
 use std::sync::Arc;
 use tokio::sync::RwLock as TokioRwLock;
-use twox_hash::XxHash64;
 use uuid::Uuid;
 
 pub struct LocalityRepositoryImpl {
@@ -65,175 +63,51 @@ impl LocalityRepositoryImpl {
 #[async_trait]
 impl LocalityRepository<Postgres> for LocalityRepositoryImpl {
     async fn save(&self, locality: LocalityModel) -> LocalityResult<LocalityModel> {
-        if !self
-            .country_subdivision_repository
-            .exists_by_id(locality.country_subdivision_id)
-            .await
-            .map_err(|e| LocalityRepositoryError::RepositoryError(e.into()))?
-        {
-            return Err(LocalityRepositoryError::CountrySubdivisionNotFound(
-                locality.country_subdivision_id,
-            ));
-        }
-
-        let query1 = sqlx::query(
-            r#"
-            INSERT INTO locality (id, country_subdivision_id, code, name_l1, name_l2, name_l3)
-            VALUES ($1, $2, $3, $4, $5, $6)
-            "#,
-        )
-        .bind(locality.id)
-        .bind(locality.country_subdivision_id)
-        .bind(locality.code.as_str())
-        .bind(locality.name_l1.as_str())
-        .bind(locality.name_l2.as_ref().map(|s| s.as_str()))
-        .bind(locality.name_l3.as_ref().map(|s| s.as_str()));
-
-        let mut hasher = twox_hash::XxHash64::with_seed(0);
-        hasher.write(locality.code.as_bytes());
-        let code_hash = hasher.finish() as i64;
-
-        let query2 = sqlx::query(
-            r#"
-            INSERT INTO locality_idx (locality_id, country_subdivision_id, code_hash)
-            VALUES ($1, $2, $3)
-            "#,
-        )
-        .bind(locality.id)
-        .bind(locality.country_subdivision_id)
-        .bind(code_hash);
-
-        match &self.executor {
-            Executor::Pool(pool) => {
-                query1
-                    .execute(&**pool)
-                    .await
-                    .map_err(|e| LocalityRepositoryError::RepositoryError(e.into()))?;
-                query2
-                    .execute(&**pool)
-                    .await
-                    .map_err(|e| LocalityRepositoryError::RepositoryError(e.into()))?;
-            }
-            Executor::Tx(tx) => {
-                let mut tx = tx.lock().await;
-                query1
-                    .execute(&mut **tx)
-                    .await
-                    .map_err(|e| LocalityRepositoryError::RepositoryError(e.into()))?;
-                query2
-                    .execute(&mut **tx)
-                    .await
-                    .map_err(|e| LocalityRepositoryError::RepositoryError(e.into()))?;
-            }
-        }
-
-        let new_idx = LocalityIdxModel {
-            locality_id: locality.id,
-            country_subdivision_id: locality.country_subdivision_id,
-            code_hash,
-        };
-        self.locality_idx_cache.read().await.add(new_idx);
-
-        Ok(locality)
+        crate::repository::person::locality_repository::save::save(self, locality).await
     }
 
     async fn load(&self, id: Uuid) -> LocalityResult<LocalityModel> {
-        let query = sqlx::query(
-            r#"
-            SELECT * FROM locality WHERE id = $1
-            "#,
-        )
-        .bind(id);
-
-        let row = match &self.executor {
-            Executor::Pool(pool) => query
-                .fetch_one(&**pool)
-                .await
-                .map_err(|e| LocalityRepositoryError::RepositoryError(e.into()))?,
-            Executor::Tx(tx) => {
-                let mut tx = tx.lock().await;
-                query
-                    .fetch_one(&mut **tx)
-                    .await
-                    .map_err(|e| LocalityRepositoryError::RepositoryError(e.into()))?
-            }
-        };
-
-        LocalityModel::try_from_row(&row)
-            .map_err(LocalityRepositoryError::RepositoryError)
+        crate::repository::person::locality_repository::load::load(self, id).await
     }
 
     async fn find_by_id(&self, id: Uuid) -> LocalityResult<Option<LocalityIdxModel>> {
-        Ok(self.locality_idx_cache.read().await.get_by_primary(&id))
+        crate::repository::person::locality_repository::find_by_id::find_by_id(self, id).await
     }
 
     async fn find_by_country_subdivision_id(
         &self,
         country_subdivision_id: Uuid,
-        _page: i32,
-        _page_size: i32,
+        page: i32,
+        page_size: i32,
     ) -> LocalityResult<Vec<LocalityIdxModel>> {
-        let cache = self.locality_idx_cache.read().await;
-        let mut result = Vec::new();
-        if let Some(ids) = cache.get_by_country_subdivision_id(&country_subdivision_id) {
-            for id in ids {
-                if let Some(idx) = cache.get_by_primary(&id) {
-                    result.push(idx);
-                }
-            }
-        }
-        Ok(result)
+        crate::repository::person::locality_repository::find_by_country_subdivision_id::find_by_country_subdivision_id(self, country_subdivision_id, page, page_size).await
     }
 
     async fn find_by_code(
         &self,
-        _country_id: Uuid,
+        country_id: Uuid,
         code: &str,
     ) -> LocalityResult<Option<LocalityIdxModel>> {
-        let mut hasher = XxHash64::with_seed(0);
-        hasher.write(code.as_bytes());
-        let code_hash = hasher.finish() as i64;
-
-        let cache = self.locality_idx_cache.read().await;
-        if let Some(id) = cache.get_by_code_hash(&code_hash) {
-            Ok(cache.get_by_primary(&id))
-        } else {
-            Ok(None)
-        }
+        crate::repository::person::locality_repository::find_by_code::find_by_code(self, country_id, code).await
     }
 
     async fn find_by_ids(&self, ids: &[Uuid]) -> LocalityResult<Vec<LocalityIdxModel>> {
-        let cache = self.locality_idx_cache.read().await;
-        let mut result = Vec::new();
-        for id in ids {
-            if let Some(idx) = cache.get_by_primary(id) {
-                result.push(idx);
-            }
-        }
-        Ok(result)
+        crate::repository::person::locality_repository::find_by_ids::find_by_ids(self, ids).await
     }
 
     async fn exists_by_id(&self, id: Uuid) -> LocalityResult<bool> {
-        Ok(self.locality_idx_cache.read().await.contains_primary(&id))
+        crate::repository::person::locality_repository::exists_by_id::exists_by_id(self, id).await
     }
 
     async fn find_ids_by_country_subdivision_id(
         &self,
         country_subdivision_id: Uuid,
     ) -> LocalityResult<Vec<Uuid>> {
-        let cache = self.locality_idx_cache.read().await;
-        Ok(cache
-            .get_by_country_subdivision_id(&country_subdivision_id)
-            .unwrap_or_default())
+        crate::repository::person::locality_repository::find_ids_by_country_subdivision_id::find_ids_by_country_subdivision_id(self, country_subdivision_id).await
     }
 
     async fn exist_by_ids(&self, ids: &[Uuid]) -> LocalityResult<Vec<bool>> {
-        let cache = self.locality_idx_cache.read().await;
-        let mut result = Vec::with_capacity(ids.len());
-        for id in ids {
-            result.push(cache.contains_primary(id));
-        }
-        Ok(result)
+        crate::repository::person::locality_repository::exist_by_ids::exist_by_ids(self, ids).await
     }
 }
