+++ banking-db-postgres/tests/suites/person/mod.rs
@@ -1,6 +1,4 @@
 // pub mod location_repository_tests;
-// pub mod locality_repository_tests;
-// pub mod locality_batch_operations_test;
 pub mod person_repository_tests;
 pub mod person_batch_operations_test;
 //pub mod location_batch_operations_test;
