+++ docs/progress/progress-tracking.md
@@ -92,6 +92,13 @@ This commit introduces batch operations for the `EntityReference` repository, si
 - **Comprehensive Testing**: A new test suite was added to ensure the correctness of all batch operations.
 - **Performance**: These changes leverage PostgreSQL's `UNNEST` for high-performance bulk inserts and updates, reducing network round-trips and database overhead.
 
+### 🎉 EntityReference Repository Decomposition and Refactoring
+This commit significantly refactors the `EntityReferenceRepositoryImpl` by decomposing its monolithic implementation into individual, single-responsibility files within a dedicated directory.
+- **Code Decomposition**: The large `entity_reference_repository_impl.rs` file has been broken down into separate files for each method (e.g., `save.rs`, `load.rs`, `find_by_id.rs`, `batch_impl.rs`), located under `repository/person/entity_reference_repository/`.
+- **Improved Modularity**: Each file now focuses on a single database operation, enhancing clarity, maintainability, and testability.
+- **Reduced File Size**: This refactoring drastically reduces the size of individual files, improving readability and LLM performance.
+- **Test Co-location**: Corresponding tests have been adjusted or co-located to align with the new modular structure.
+
 ### 🎉 Person Domain Refactoring: Batch Operations and Caching
 This commit introduces a significant refactoring of the Person domain, focusing on optimizing batch operations and improving data consistency with transaction-aware caching.
 - **Batch Operations**: The `person_repository_batch_impl` has been heavily refactored for performance and clarity when handling bulk data operations.
